1.0.0
Hi, I'm just passing the time. I hope the framework I made can be a learning material for my github visitors. I'm not sure if it's called a framework, I feel like it's a biolerplate or micro-framework, but feel free to respond.

I hereby release the framework from version 1

Thanks.